Combined the checks in logprob and the corresponding tests suites
